did some updates in book services and order services

+++ All-Capg-Workspaces/Capg-SB-MicroServices-Workspace/BookStore-Microservices/Assignment_14_BookService/src/main/java/com/capg/bookservice/controller/BookServiceController.java
@@ -34,7 +34,7 @@ public class BookServiceController {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Book> getBookById(@PathVariable("id") int id) {
+	public ResponseEntity<Book> getBookById(@PathVariable("id") long id) {
 		
 		Book book = service.getBookById(id);
